@@ -1,23 +1,30 @@
 {
-    "name": "nocksapp/tivoka",
-    "type": "library",
-    "description": "The universal JSON-RPC client/server library. JSON-RPC done right!",
-    "keywords": ["json", "rpc", "json-rpc", "api", "xml-rpc"],
-    "homepage": "http://marcelklehr.github.com/tivoka",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "Marcel Klehr",
-            "email": "mklehr@gmx.net"
-        }
-    ],
-    "require": {
-        "php": ">=5.3",
-        "textalk/websocket": "1.0.*"
-    },
-    "autoload": {
-        "psr-0" : {
-            "Tivoka" : "lib/"
-        }
+  "name": "bug-dealer/tivoka",
+  "version": "4.0.2",
+  "type": "library",
+  "description": "The universal JSON-RPC client/server library. JSON-RPC done right!",
+  "keywords": [
+    "json",
+    "rpc",
+    "json-rpc",
+    "api",
+    "xml-rpc"
+  ],
+  "homepage": "https://github.com/bug-dealer/tivoka",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Marcel Klehr",
+      "email": "mklehr@gmx.net"
     }
+  ],
+  "require": {
+    "php": ">=7.3",
+    "textalk/websocket": "1.0.*"
+  },
+  "autoload": {
+    "psr-0": {
+      "Tivoka": "lib/"
+    }
+  }
 }
