@@ -1,16 +1,16 @@
 <?php
-include('../include.php');
 
-$methods = array(
-	'demo.sayHello' => function() {
-		return 'Hello World!';
-	},
-	
-	'demo.substract' => function($params) {
-		list($num1, $num2) = $params;
-		return $num1 - $num2;
-	}
-);
+include(__DIR__ . '../include.php');
+$methods = [
+    'demo.sayHello' => function (): string {
 
+        return 'Hello World!';
+    },
+
+    'demo.substract' => function ($params) {
+
+        [$num1, $num2] = $params;
+        return $num1 - $num2;
+    }
+];
 Tivoka\Server::provide($methods)->dispatch();
-?>
