@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Tivoka - JSON-RPC done right!
  * Copyright (c) 2011-2012 by Marcel Klehr <mklehr@gmx.net>
@@ -42,11 +43,10 @@ class Notification extends Request
      * @param mixed $params Additional params for the remote procedure
      * @see Tivoka_Connection::send()
      */
-    public function __construct($method, $params=null)
+    public function __construct($method, $params = null)
     {
         $this->id = null;
         $this->method = $method;
         $this->params = $params;
     }
 }
-?>
